requirements.txt
absl-py
imgaug
tensorflow==2.1.0rc1
opencv-python==4.1.1.26
lxml
tqdm
